requirements.txt
streamlit>=1.33
numpy>=1.24
pandas>=2.0
networkx>=3.1
scipy>=1.10
plotly>=5.18
openpyxl>=3.1
joblib>=1.3
pyyaml>=6.0
pytest>=7.4

kaleido>=0.2
